11-DRF-user-roles/requirements.txt
Django==1.11.6
django-debug-toolbar==1.8
django-request-logging==0.5.3
djangorestframework==3.7.1
mysqlclient==1.3.12
Pygments==2.2.0
pytz==2017.2
sqlparse==0.2.4